Create CarbonRevenueAccount.sol
A sample 'carbon fee and dividend' revenue account.  The concept is very loosely based on the Energy Innovation and Carbon Dividend Act of 2019 (H.R. 763).  The idea is that a portion of fees collected for carbon intensive goods and services can be refunded back to the population as a equitable dividend.